area of circle

+++ exercises/area_of_circle.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
 
-int main(){
+int main()
+{
 
-printf("Welcome! Please enter a radius\n");
-int radius;
-scanf("%i", &radius);
-int area = 3.14 * (radius * radius);
-printf("The area of the circle is %i given the radius %i\n", area, radius);
+    printf("Welcome! Please enter a radius\n");
+    int radius;
+    scanf("%i", &radius);
+    double area = 3.14 * (radius * radius);
+    printf("The area of the circle is %f given the radius %i\n", area, radius);
     return 0;
 }
